Integração

+++ back-end/ledes/src/main/resources/db/migration/V13_CriacaoRelacionamentoMembroPapel.sql
@@ -1,6 +1,6 @@
 CREATE TABLE membro_tipo_papel (
-    membro_id BIGINT,
-    tipo_papel_id BIGINT,
+    membro_id INT,
+    tipo_papel_id INT,
     PRIMARY KEY (membro_id, tipo_papel_id),
     FOREIGN KEY (membro_id) REFERENCES membro(id),
     FOREIGN KEY (tipo_papel_id) REFERENCES tipo_papel(id)
