+++ back-end/ledes/src/main/resources/db/migration/V14_CriacaorelacionamentoMembroVinculo.sql
@@ -1,6 +1,6 @@
 CREATE TABLE membro_tipo_vinculo (
-    membro_id BIGINT,
-    tipo_vinculo_id BIGINT,
+    membro_id INT,
+    tipo_vinculo_id INT,
     PRIMARY KEY (membro_id, tipo_vinculo_id),
     FOREIGN KEY (membro_id) REFERENCES membro(id),
     FOREIGN KEY (tipo_vinculo_id) REFERENCES tipo_vinculo(id)
